move page_view functions to lib
