--- src/auth_failure.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="jp">
-
-<head>
-  <meta charset="utf-8">
-  <title>
-    Auth_Failure
-  </title>
-</head>
-
-<body>
-  <div class="container">
-    <h1>Welcome {{ name }}!</h1>
-    <h1>Your Session ID is  " {{ UserId }} "</h1>
-    <!-- <p><a href="book.html">Book</a></p> -->
-    <p><a href="http://localhost:8080/app/book">Book</a></p>
-    <p><a href="http://localhost:8080/app/clothing">Clothing</a></p>
-  </div>
-</body>
-
-</html>
